+++ src/BlitzRecommendations.php
@@ -11,9 +11,11 @@ use craft\elements\db\ElementQuery;
 use craft\events\CancelableEvent;
 use craft\events\RegisterComponentTypesEvent;
 use craft\services\Utilities;
+use craft\web\Response;
 use putyourlightson\blitzrecommendations\services\RecommendationsService;
 use putyourlightson\blitzrecommendations\utilities\RecommendationsUtility;
 use yii\base\Event;
+use yii\web\Response as BaseResponse;
 
 /**
  * @property RecommendationsService $recommendations
@@ -23,7 +25,24 @@ class BlitzRecommendations extends Plugin
     /**
      * @var BlitzRecommendations
      */
-    public static $plugin;
+    public static BlitzRecommendations $plugin;
+
+    /**
+     * @inheritdoc
+     */
+    public static function config(): array
+    {
+        return [
+            'components' => [
+                'recommendations' => ['class' => RecommendationsService::class],
+            ],
+        ];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public string $schemaVersion = '2.0.0';
 
     /**
      * @inheritdoc
@@ -31,12 +50,7 @@ class BlitzRecommendations extends Plugin
     public function init()
     {
         parent::init();
-
         self::$plugin = $this;
-
-        $this->setComponents([
-            'recommendations' => RecommendationsService::class,
-        ]);
 
         $this->_registerEvents();
         $this->_registerUtilities();
@@ -47,17 +61,26 @@ class BlitzRecommendations extends Plugin
      */
     private function _registerEvents()
     {
+        // Ignore CP requests
+        if (Craft::$app->getRequest()->getIsCpRequest()) {
+            return;
+        }
+
         // Register element query prepare event
         Event::on(ElementQuery::class, ElementQuery::EVENT_BEFORE_PREPARE,
             function(CancelableEvent $event) {
-                // Ignore CP requests
-                if (Craft::$app->getRequest()->getIsCpRequest()) {
-                    return;
-                }
-
                 /** @var ElementQuery $elementQuery */
                 $elementQuery = $event->sender;
                 $this->recommendations->checkElementQuery($elementQuery);
+            },
+            null,
+            false
+        );
+
+        // Register element query prepare event
+        Event::on(Response::class, BaseResponse::EVENT_AFTER_PREPARE,
+            function() {
+                $this->recommendations->save();
             },
             null,
             false