+++ src/records/RecommendationRecord.php
@@ -12,19 +12,15 @@ use DateTime;
  * @property int $id
  * @property string $key
  * @property string $template
+ * @property string $line
  * @property string $message
  * @property string $info
  * @property DateTime $lastUpdated
  */
 class RecommendationRecord extends ActiveRecord
 {
-    // Public Static Methods
-    // =========================================================================
-
     /**
      * @inheritdoc
-     *
-     * @return string
      */
     public static function tableName(): string
     {